render: properly defined "out of bounds time" behavior with footage
If a user extends a clip beyond the footage's length, they can
choose whether they want a black frame, the last frame continued,
or for the footage to loop infinitely (same for extending earlier
than the start of footage).

+++ app/codec/waveinput.h
@@ -40,9 +40,9 @@ public:
 
   bool is_open() const;
 
-  QByteArray read(int length);
-  QByteArray read(int offset, int length);
-  qint64 read(int offset, char *buffer, int length);
+  QByteArray read(qint64 length);
+  QByteArray read(qint64 offset, qint64 length);
+  qint64 read(qint64 offset, char *buffer, qint64 length);
 
   bool seek(qint64 pos);
